backend/requirements.txt
# =============================================================================
# EcoQuant Backend - Production-Grade Quant Investment Platform
# =============================================================================
# Python 3.11+ Required

# -----------------------------------------------------------------------------
# Web Framework & ASGI
# -----------------------------------------------------------------------------
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# -----------------------------------------------------------------------------
# Database (Async PostgreSQL + SQLAlchemy 2.0)
# -----------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Sync driver for Celery workers
alembic==1.13.1

# -----------------------------------------------------------------------------
# Validation & Serialization (Pydantic V2)
# -----------------------------------------------------------------------------
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# -----------------------------------------------------------------------------
# Task Queue (Celery + Redis)
# -----------------------------------------------------------------------------
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1  # Celery monitoring

# -----------------------------------------------------------------------------
# Quant & Data Analysis
# -----------------------------------------------------------------------------
backtrader==1.9.78.123
pandas==2.2.0
numpy==1.26.4
yfinance==0.2.36
ta-lib==0.4.28  # Technical Analysis (optional, requires system lib)

# -----------------------------------------------------------------------------
# AI Services
# -----------------------------------------------------------------------------
google-generativeai==0.8.3  # Google Gemini API for news sentiment analysis

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
httpx==0.26.0
python-dotenv==1.0.1
structlog==24.1.0  # Structured logging

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.0
mypy==1.8.0

